+++ app/src/main/res/layout/fragment_end_game.xml
@@ -5,7 +5,7 @@
     android:id="@+id/fragment_end_game"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".EndGame">
+    tools:context=".Fragments.EndGameFragment">
 
     <GridLayout
         android:id="@+id/EGHor1"
