+++ app/src/main/res/layout/fragment_pause_menu.xml
@@ -5,7 +5,7 @@
     android:id="@+id/fragment_pause_menu"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PauseMenu">
+    tools:context=".Fragments.PauseMenuFragment">
 
     <GridLayout
         android:id="@+id/PauseMenuHor1"
