+++ app/src/main/res/layout/fragment_select_game_mode.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".GameModeSelectFragment">
+    tools:context=".Fragments.GameModeSelectFragment">
 
 
     <LinearLayout
