+++ app/src/main/res/layout/fragment_select_player_to_customize.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SelectPlayerToCustomizeFragment">
+    tools:context=".Fragments.SelectPlayerToCustomizeFragment">
 
     <androidx.constraintlayout.widget.Guideline
         android:id="@+id/side_bar_guideline"
