+++ app/src/main/res/layout/fragment_small_game_board.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/design_default_color_secondary_variant"
-    tools:context=".SmallGameBoardFragment">
+    tools:context=".Fragments.SmallGameBoardFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
